Сообщения загружаются по запросу

+++ KfkAdmin/Components/Pages/Home/Components/BrokerTable.razor.cs
@@ -20,13 +20,13 @@ public partial class BrokerTable(IKafkaRepositoryProvider repositoryProvider) : 
             Port = x.Port,
             PartitionCount = partitions.Count(p => p.BrokerId == x.BrokerId),
             PartitionPercent = partitions.Count(p => p.BrokerId == x.BrokerId) > 0 ? 
-                (double)partitions.Count(p => p.BrokerId == x.BrokerId) / partitions.Count * 100 : 0
+                (decimal)partitions.Count(p => p.BrokerId == x.BrokerId) / partitions.Count * 100 : 0
         }).ToList();
     }
     
     private class BrokerTableViewModel : Broker
     {
         public int PartitionCount { get; set; }
-        public double PartitionPercent { get; set; }
+        public decimal PartitionPercent { get; set; }
     }
 }
